fix bench observation audit drift
